+++ CHANGELOG.md
@@ -25,3 +25,4 @@
 - [feat:Add GitHub action of codecov.yml.](https://github.com/Tencent/spring-cloud-tencent/pull/328)
 - [Feature: add spring cloud tencent logo](https://github.com/Tencent/spring-cloud-tencent/pull/329)
 - [Feature: Optimize static metadata manager](https://github.com/Tencent/spring-cloud-tencent/pull/327)
+- [test:update junit of metadata.](https://github.com/Tencent/spring-cloud-tencent/pull/340)